f2py: move Fortran utilities into codegen/utils and handle BIND(C)
Merge clu/fortran into codegen/utils alongside gen_forward_decls; add
linkage_name and is_fortran_value to the IR and resolve BIND(C) symbol names.
